Moved date interval writing funciton to library and fixed it
